requirements.txt

Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
# Ajoutez ces lignes si vous utilisez PostgreSQL ou MySQL:
# psycopg2-binary==2.9.9  # Pour PostgreSQL
# PyMySQL==1.1.0          # Pour MySQL
SQLAlchemy==2.0.44
APScheduler==3.11.1
icalendar==6.3.2
requests==2.32.5
cryptography==46.0.3
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.1.0
pycparser==2.23
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
